wip: splitting API logic to be more readable, finished for Price

+++ api-dotnet/CryptoTracker.Api/Models/Price.cs
@@ -11,5 +11,5 @@ public class Price
     public DateTimeOffset Timestamp { get; set; }
 
     [Column("price")]
-    public decimal PriceValue { get; set; }
+    public double PriceValue { get; set; }
 }
